Fix flatten with pointers containing escaped chars (e.g. '{}'
Fixed undue duplicate name detection when resolving remote refs

Improves behavior in edge cases where schemas points to responses
or parameters.

This makes the flatten behavior consistent with swagger validate.

* now unmarshals the struct as a schema: if this works, flattening
moves on, with a warning to the user (as this is valid but most
likely unintended)
* improves go-swagger/go-swagger#1602 and go-swagger/go-swagger#1614
* contributes go-swagger/go-swagger#1621
* Expand schemas when $ref'ed to (shared) responses or parameters

* Added internal pre go1.8 PathUnescape (like in go-openapi/swag)
* Added test hack for coverage with multiple packages

Added test config in CI for package internal

* Removed CI test with go 1.7
  This commit remains compatible with go1.7, but some dependencies
  (i.e. github.com/asaskevich/govalidator) are no more compatible
  using master